awssqs: consolidate pkg name
Following the same standard as used for all other sources under
pkg/controller.
