+++ pkg/controller/awssqssource/resources/receive_adapter.go
